Add Kafka and Minio
