+++ langstream-cli/src/main/java/ai/langstream/cli/commands/applications/AbstractDeployApplicationCmd.java
@@ -15,32 +15,16 @@
  */
 package ai.langstream.cli.commands.applications;
 
-import ai.langstream.admin.client.HttpRequestFailedException;
-import ai.langstream.admin.client.http.HttpClientFacade;
 import ai.langstream.admin.client.util.MultiPartBodyPublisher;
 import ai.langstream.cli.commands.GitIgnoreParser;
 import ai.langstream.cli.util.LocalFileReferenceResolver;
-import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.IOException;
-import java.net.URI;
-import java.net.URL;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.security.DigestInputStream;
-import java.security.MessageDigest;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.SneakyThrows;
 import net.lingala.zip4j.ZipFile;
 import net.lingala.zip4j.model.ZipParameters;
@@ -95,11 +79,6 @@ public abstract class AbstractDeployApplicationCmd extends BaseApplicationCmd {
         boolean isUpdate() {
             return false;
         }
-
-        @Override
-        boolean isLocalRun() {
-            return false;
-        }
     }
 
     @CommandLine.Command(name = "update", header = "Update an existing LangStream application")
@@ -147,64 +126,6 @@ public abstract class AbstractDeployApplicationCmd extends BaseApplicationCmd {
         boolean isUpdate() {
             return true;
         }
-
-        @Override
-        boolean isLocalRun() {
-            return false;
-        }
-    }
-
-    @CommandLine.Command(
-            name = "local-run",
-            header = "Run on a docker container a LangStream application")
-    public static class LocalRun extends AbstractDeployApplicationCmd {
-
-        @CommandLine.Option(
-                names = {"-app", "--application"},
-                description = "Application directory path",
-                required = true)
-        private String appPath;
-
-        @CommandLine.Option(
-                names = {"-i", "--instance"},
-                description = "Instance file path",
-                required = true)
-        private String instanceFilePath;
-
-        @CommandLine.Option(
-                names = {"-s", "--secrets"},
-                description = "Secrets file path")
-        private String secretFilePath;
-
-        @Override
-        String applicationId() {
-            return "app";
-        }
-
-        @Override
-        String appPath() {
-            return appPath;
-        }
-
-        @Override
-        String instanceFilePath() {
-            return instanceFilePath;
-        }
-
-        @Override
-        String secretFilePath() {
-            return secretFilePath;
-        }
-
-        @Override
-        boolean isUpdate() {
-            return false;
-        }
-
-        @Override
-        boolean isLocalRun() {
-            return true;
-        }
     }
 
     abstract String applicationId();
@@ -216,8 +137,6 @@ public abstract class AbstractDeployApplicationCmd extends BaseApplicationCmd {
     abstract String secretFilePath();
 
     abstract boolean isUpdate();
-
-    abstract boolean isLocalRun();
 
     @Override
     @SneakyThrows
@@ -277,152 +196,15 @@ public abstract class AbstractDeployApplicationCmd extends BaseApplicationCmd {
             }
         }
 
-        if (isLocalRun()) {
-            executeOnDocker(appDirectory, instanceContents, secretsContents);
+        final MultiPartBodyPublisher bodyPublisher = buildMultipartContentForAppZip(contents);
+
+        if (isUpdate()) {
+            getClient().applications().update(applicationId, bodyPublisher);
+            log(String.format("application %s updated", applicationId));
         } else {
-            final MultiPartBodyPublisher bodyPublisher = buildMultipartContentForAppZip(contents);
-
-            if (isUpdate()) {
-                getClient().applications().update(applicationId, bodyPublisher);
-                log(String.format("application %s updated", applicationId));
-            } else {
-                getClient().applications().deploy(applicationId, bodyPublisher);
-                log(String.format("application %s deployed", applicationId));
-            }
+            getClient().applications().deploy(applicationId, bodyPublisher);
+            log(String.format("application %s deployed", applicationId));
         }
-    }
-
-    private void executeOnDocker(File appDirectory, String instanceContents, String secretsContents)
-            throws Exception {
-        File tmpInstanceFile = Files.createTempFile("instance", ".yaml").toFile();
-        Files.write(tmpInstanceFile.toPath(), instanceContents.getBytes(StandardCharsets.UTF_8));
-        File tmpSecretsFile = Files.createTempFile("secrets", ".yaml").toFile();
-        Files.write(tmpSecretsFile.toPath(), secretsContents.getBytes(StandardCharsets.UTF_8));
-        String imageName = "langstream/langstream-runtime-tester:latest-dev";
-        List<String> commandLine = new ArrayList<>();
-        commandLine.add("docker");
-        commandLine.add("run");
-        commandLine.add("--rm");
-        commandLine.add("-it");
-        commandLine.add("-v");
-        commandLine.add(appDirectory.getAbsolutePath() + ":/code/application");
-        commandLine.add("-v");
-        commandLine.add(tmpInstanceFile.getAbsolutePath() + ":/code/instance.yaml");
-        commandLine.add("-v");
-        commandLine.add(tmpSecretsFile.getAbsolutePath() + ":/code/secrets.yaml");
-        commandLine.add(imageName);
-        ProcessBuilder processBuilder = new ProcessBuilder(commandLine).inheritIO();
-        Process process = processBuilder.start();
-        process.waitFor();
-    }
-
-    @AllArgsConstructor
-    @Getter
-    private static class Dependency {
-        private String name;
-        private String url;
-        private String type;
-        private String sha512sum;
-    }
-
-    private void downloadDependencies(Path directory) throws Exception {
-
-        final Path configuration = directory.resolve("configuration.yaml");
-        if (!Files.exists(configuration)) {
-            return;
-        }
-        final Map<String, Object> map =
-                yamlConfigReader.readValue(configuration.toFile(), Map.class);
-        final List<Map<String, Object>> dependencies =
-                (List<Map<String, Object>>) map.get("dependencies");
-        if (dependencies == null) {
-            return;
-        }
-        final List<Dependency> dependencyList =
-                dependencies.stream()
-                        .map(
-                                dependency ->
-                                        new Dependency(
-                                                (String) dependency.get("name"),
-                                                (String) dependency.get("url"),
-                                                (String) dependency.get("type"),
-                                                (String) dependency.get("sha512sum")))
-                        .collect(Collectors.toList());
-
-        for (Dependency dependency : dependencyList) {
-            URL url = new URL(dependency.getUrl());
-
-            final String outputPath;
-            switch (dependency.getType()) {
-                case "java-library":
-                    outputPath = "java/lib";
-                    break;
-                default:
-                    throw new RuntimeException(
-                            "unsupported dependency type: " + dependency.getType());
-            }
-
-            Path output = directory.resolve(outputPath);
-            if (!Files.exists(output)) {
-                Files.createDirectories(output);
-            }
-            String rawFileName = url.getFile().substring(url.getFile().lastIndexOf('/') + 1);
-            Path fileName = output.resolve(rawFileName);
-
-            if (Files.isRegularFile(fileName)) {
-
-                if (!checkChecksum(fileName, dependency.getSha512sum())) {
-                    log("File seems corrupted, deleting it");
-                    Files.delete(fileName);
-                } else {
-                    log(String.format("Dependency: %s at %s", fileName, fileName.toAbsolutePath()));
-                    continue;
-                }
-            }
-
-            log(
-                    String.format(
-                            "downloading dependency: %s to %s",
-                            fileName, fileName.toAbsolutePath()));
-            final HttpRequest request = getClient().newDependencyGet(url);
-            getClient().http(request, HttpResponse.BodyHandlers.ofFile(fileName));
-
-            if (!checkChecksum(fileName, dependency.getSha512sum())) {
-                log("File still seems corrupted. Please double check the checksum and try again.");
-                Files.delete(fileName);
-                throw new IOException("File at " + url + ", seems corrupted");
-            }
-
-            log("dependency downloaded");
-        }
-    }
-
-    private boolean checkChecksum(Path fileName, String sha512sum) throws Exception {
-        MessageDigest instance = MessageDigest.getInstance("SHA-512");
-        try (DigestInputStream inputStream =
-                new DigestInputStream(
-                        new BufferedInputStream(Files.newInputStream(fileName)), instance)) {
-            while (inputStream.read() != -1) {}
-        }
-        byte[] digest = instance.digest();
-        String base16encoded = bytesToHex(digest);
-        if (!sha512sum.equals(base16encoded)) {
-            log(String.format("Computed checksum: %s", base16encoded));
-            log(String.format("Expected checksum: %s", sha512sum));
-        }
-        return sha512sum.equals(base16encoded);
-    }
-
-    private static String bytesToHex(byte[] hash) {
-        StringBuilder hexString = new StringBuilder(2 * hash.length);
-        for (byte b : hash) {
-            String hex = Integer.toHexString(0xff & b);
-            if (hex.length() == 1) {
-                hexString.append('0');
-            }
-            hexString.append(hex);
-        }
-        return hexString.toString();
     }
 
     public static Path buildZip(File appDirectory, Consumer<String> logger) throws IOException {
@@ -474,64 +256,6 @@ public abstract class AbstractDeployApplicationCmd extends BaseApplicationCmd {
                 }
             }
         }
-    }
-
-    @SneakyThrows
-    private File checkFileExistsOrDownload(String path) {
-        if (path == null) {
-            return null;
-        }
-        if (path.startsWith("http://")) {
-            throw new IllegalArgumentException("http is not supported. Please use https instead.");
-        }
-        if (path.startsWith("https://")) {
-            return downloadHttpsFile(path, getClient().getHttpClientFacade(), this::log);
-        }
-        if (path.startsWith("file://")) {
-            path = path.substring("file://".length());
-        }
-        final File file = new File(path);
-        if (!file.exists()) {
-            throw new RuntimeException("File " + path + " does not exist");
-        }
-        return file;
-    }
-
-    static File downloadHttpsFile(String path, HttpClientFacade client, Consumer<String> logger)
-            throws IOException, HttpRequestFailedException {
-        final URI uri = URI.create(path);
-        if ("github.com".equals(uri.getHost())) {
-            return downloadFromGithub(uri, logger);
-        }
-
-        final HttpRequest request =
-                HttpRequest.newBuilder()
-                        .uri(URI.create(path))
-                        .version(HttpClient.Version.HTTP_1_1)
-                        .GET()
-                        .build();
-        final Path tempFile = Files.createTempFile("langstream", ".bin");
-        final long start = System.currentTimeMillis();
-        // use HttpResponse.BodyHandlers.ofByteArray() to get cleaner error messages
-        final HttpResponse<byte[]> response =
-                client.http(request, HttpResponse.BodyHandlers.ofByteArray());
-        if (response.statusCode() >= 400) {
-            throw new RuntimeException(
-                    "Failed to download file: "
-                            + path
-                            + "\nReceived status code: "
-                            + response.statusCode()
-                            + "\n"
-                            + response.body());
-        }
-        Files.write(tempFile, response.body());
-        final long time = (System.currentTimeMillis() - start) / 1000;
-        logger.accept(String.format("downloaded remote file %s (%d s)", path, time));
-        return tempFile.toFile();
-    }
-
-    private static File downloadFromGithub(URI uri, Consumer<String> logger) {
-        return GithubRepositoryDownloader.downloadGithubRepository(uri, logger);
     }
 
     public static MultiPartBodyPublisher buildMultipartContentForAppZip(
